NPC tooltips now include icon
Added small icons to the tooltips for abilities and items so that users can quickly see if the tooltip corresponds to the correct item or ability. This required parsing additional files from the game files, uploading them to a Google Cloud Bucket, and referring to those images in the client.