auto-reconnect to server without losing running jobs
Added support for the server to maintain currently running jobs when the client disconnects, up to some timeout (now defined in the options as SOCKET_TIMEOUT). This also requires the websocket client to store a session ID and send to the server when it is known, or to request a new one when it is starting up for the first time (e.g. on install or when chrome closes and reopens).