new style for connection status
Changed the way connection to the server is handled. The two major differences are (1) added a "connection status" indicator led to the monitor, and (2) took away intermediate screens showing "Connecting" and disconnection errors. Now, when the user clicks Start, he is always taken to the tracker page and text on the tracker page and the indicator LED indicate the connection status instead of having to wait for the connection to occur. A major benefit of this is that users can view saved content without needing to be connected to the server (ever). This also required or resulted in  several other changes:
- popup and background now store a string "connection" status instead of boolean "connected"
- popup now sets the LED and other connection dependent content when the background page updates the connection status
- fixed a bug in the background page where if the user was diconnected and tried to reconnect, the background page wouldnt do anything
- added the LED to popup.html
- removed all the intermediate pages in the popup, i.e. the connecting page, disconnected, reconnecting pages
- gathered a bunch of the popup updates into fewer (or single) functions
- removed the Connecting animation stuff from the style.css
- when disconnected or connecting, the tracker page now disables Go Online and make a new request links (which it was supposed to do before but didnt work)
- fixed a bug in jobs.py where direct downloads of replays (e.g. in the sample videos) were handled wrong (was not correctly assessing if a file was compressed or not)