feature added: no replays required
No replays are required for the user to submit a video. This removes all match information from replays and lists only default abilities. So, e.g. items are not shown, stolen or invoked abilities are not shown, other info not shown. Adding this feature required the following changes:
- parse_resources now saves to the server's keys.pkl a default ordering and ID of hero abilities
- jobs.py now accepts None everywhere where replays were formerly not optional
- jobs.py fills in hero info with default info for that hero's abilities when no replay data are available
- jobs_tornado.py now has replays as an optional argument

Other changes:
- background now asks the server to kill all running jobs when the extension is unloaded (e.g. browser closed or extension reloaded). This is to prevent the server from being stuck with zombie jobs.
- fixed the call to chrome.storage.onChanged to refer to the correct storage location ('sync' instead of "storage")